+++ backend_core/tomato/lib/remote_info.py
@@ -1,0 +1,1 @@
+../../../shared/lib/remote_info.py
